* fixed rare bug with storage, also now when setting it to -1 it doesn't erases key from map * formating * added multi(-client)checking lua command
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@205 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ iologindata.cpp
@@ -825,9 +825,9 @@ bool IOLoginData::savePlayer(Player* player, bool preSave)
 
 	query_insert.setQuery("INSERT INTO `player_storage` (`player_id`, `key`, `value`) VALUES ");
 	player->genReservedStorageRange();
-	for(StorageMap::const_iterator cit = player->getStorageIteratorBegin(); cit != player->getStorageIteratorEnd();cit++)
+	for(StorageMap::const_iterator cit = player->getStorageIteratorBegin(); cit != player->getStorageIteratorEnd(); cit++)
 	{
-		sprintf(buffer, "(%d, %d, %u)", player->getGUID(), cit->first, (uint32_t)cit->second);
+		sprintf(buffer, "(%u, %u, %d)", player->getGUID(), cit->first, cit->second);
 		if(!query_insert.addRow(buffer))
 			return false;
 	}
